somewhat refactored ITempBuildings and CurrentBuildingData types, changed delete temp to a boolean, not a complete deletion from db for now
